Update app.rs, notifications.rs, settings.rs, and 38 more files
Implement more piecemeal parsing functions for the local data structures, to minimize data loss if/when the data which has been stored to file fails to parse, due to data structure changes/etc.
Added serde default annotations to the various data structures.
Normalized all uses of usize to u64.
Normalized string parameter types in io.rs to String.
Updated App::initializeState to downgrade to the lossy parsing methods when the user data fails to parse normally.
Added logging statements to App::initializeState.

+++ src/retroachievements/components/achievement.rs
@@ -11,7 +11,7 @@ use crate::retroachievements::platform::api::Api;
 use crate::{join, png, pngAlt, RetroAchievementsUserData};
 
 #[component]
-pub fn AchievementElement(gameId: usize, achievementId: usize) -> Element
+pub fn AchievementElement(gameId: u64, achievementId: u64) -> Element
 {
 	let game = match RetroAchievementsUserData().games.iter()
 		.find(|g| g.id == gameId)
